requirements.txt
h5py >=2.9
numpy >= 1.19
Theano == 0.8.2
biopython >= 1.61
progressbar33 >= 2.3.1
six >= 1.10
ont-fast5-api